+++ lib/flutter_miui.dart
@@ -8,3 +8,4 @@ export 'src/miui_button.dart';
 export 'src/miui_menu.dart';
 export 'src/miui_global.dart';
 export 'src/miui_textfield.dart';
+export 'src/animated_text.dart';